requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
boto3==1.42.90
botocore==1.42.90
click==8.3.2
fastapi==0.136.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
